% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.nnet-package}
\alias{mlr3learners.nnet}
\alias{mlr3learners.nnet-package}
\title{mlr3learners.nnet: nnet and multinom Learner for mlr3}
\description{
nnet Learner for mlr3.
}
\author{
\strong{Maintainer}: Henri Funk \email{henrifnk@gmail.com}

}
